+++ src/curves.jl
@@ -69,7 +69,7 @@ function center3pts(p1::Point, p2::Point, p3::Point)
         r = √(norm2(p1-c))
         return c, r
     else
-        @warn "There are no circles which pass through $p1, $p2 and $p3."
+        @warn "center3pts(): There are no circles which pass through $p1, $p2 and $p3."
         return (Point(0, 0), 0)
     end
 end